@@ -63,8 +63,7 @@ export default class MySQLEndpoint extends AEndpointStream
 
   public async getSource(source: SourceConfig): Promise<Readable>
   {
-    const integrationId = source.options['integrationId'];
-    const config: MySQLConfig = await this.getConfig(integrationId, source.options);
+    const config: MySQLConfig = await this.getIntegrationConfig(source.integrationId);
     const table = source.options['table'];
     const query: string = source.options['query'];
     return new MySQLReader(config, query, table);
@@ -72,38 +71,11 @@ export default class MySQLEndpoint extends AEndpointStream
 
   public async getSink(sink: SinkConfig, engine?: TransformationEngine): Promise<Writable>
   {
-    const integrationId = sink.options['integrationId'];
-    const config: MySQLConfig = await this.getConfig(integrationId, sink.options);
+    const config: MySQLConfig = await this.getIntegrationConfig(sink.integrationId);
     const table = sink.options['table'];
     const query: string = sink.options['query'];
 
     throw new Error('not implemented');
     // return new MySQLWriter(config, database, table);
-  }
-
-  private async getConfig(integrationId: number, options?: object): Promise<MySQLConfig>
-  {
-    return new Promise(async (resolve, reject) =>
-    {
-      const integration: IntegrationConfig[] = await integrations.get(null, integrationId);
-      if (integration.length === 0)
-      {
-        return reject(new Error('Invalid MySQL integration ID.'));
-      }
-
-      let mysqlConfig: MySQLConfig = {};
-      try
-      {
-        const connectionConfig = JSON.parse(integration[0].connectionConfig);
-        const authConfig = JSON.parse(integration[0].authConfig);
-        mysqlConfig = Object.assign(connectionConfig, authConfig);
-      }
-      catch (e)
-      {
-        return reject(new Error('Retrieving integration ID ' + String(integrationId)));
-      }
-
-      resolve(mysqlConfig);
-    });
   }
 }
