@@ -55,6 +55,7 @@ import * as Tasty from '../../tasty/Tasty';
 import * as App from '../App';
 import { Job } from './Job';
 import JobConfig from './JobConfig';
+import JobLogConfig from './JobLogConfig';
 import { Task } from './Task';
 import { TaskTree } from './TaskTree';
 
@@ -80,6 +81,7 @@ export class JobQueue
       [
         'createdAt',
         'createdBy',
+        'endTime',
         'logId',
         'meta',
         'name',
@@ -88,6 +90,7 @@ export class JobQueue
         'running',
         'runNowPriority',
         'scheduleId',
+        'startTime',
         'status',
         'tasks',
         'type',
@@ -111,7 +114,7 @@ export class JobQueue
           this.runningJobs.delete(id);
 
           // set status to CANCELED
-          await this._setJobStatus(id, false, 'CANCELED');
+          await this.setJobStatus(id, false, 'CANCELED');
           return resolve(await this.get(id) as JobConfig[]);
         }
         else if (this.runningRunNowJobs.get(id) !== undefined)
@@ -123,7 +126,7 @@ export class JobQueue
           this.runningRunNowJobs.delete(id);
 
           // set status to CANCELED
-          await this._setJobStatus(id, false, 'CANCELED');
+          await this.setJobStatus(id, false, 'CANCELED');
           return resolve(await this.get(id) as JobConfig[]);
         }
       }
@@ -161,13 +164,13 @@ export class JobQueue
         return reject(new Error('Job ID must not be provided as it is autogenerated'));
       }
       // != undefined is a way to include both undefined and null
-      const creationDate: Date = new Date();
-      job.createdAt = creationDate;
+      job.createdAt = new Date();
       if (job.id === null)
       {
         delete job.id;
       }
       job.createdBy = userId !== undefined ? userId : null;
+      job.endTime = null;
       job.meta = (job.meta !== undefined && job.meta !== null) ? job.meta : '';
       job.name = (job.name !== undefined && job.name !== null) ? job.name : '';
       job.pausedFilename = (job.pausedFilename !== undefined && job.pausedFilename !== null) ? job.pausedFilename : '';
@@ -175,6 +178,7 @@ export class JobQueue
       job.running = (job.running !== undefined && job.running !== null) ? job.running : false;
       job.runNowPriority = (job.runNowPriority !== undefined && job.runNowPriority !== null) ? job.runNowPriority : 1;
       job.scheduleId = (job.scheduleId !== undefined) ? job.scheduleId : null;
+      job.startTime = null;
       job.status = (job.status !== undefined && job.status !== null && job.status !== '') ? job.status : 'PENDING';
       job.tasks = (job.tasks !== undefined && job.tasks !== null) ? job.tasks : '[]';
       job.type = (job.type !== undefined && job.type !== null) ? job.type : 'default';
@@ -211,9 +215,26 @@ export class JobQueue
     return this._select([], { id, running });
   }
 
-  public async getLog(id: number): Promise<object>
+  public async getLog(id?: number): Promise<JobLogConfig>
   {
-    return Promise.resolve({}); // TODO implement this
+    return new Promise<JobLogConfig>(async (resolve, reject) =>
+    {
+      const jobs: JobConfig[] = await this.get(id);
+      if (jobs.length === 0)
+      {
+        throw new Error('Job not found.');
+      }
+      const jobLogConfig: JobLogConfig[] = await App.JobL.get(id);
+      return resolve(jobLogConfig[0]);
+    });
+  }
+
+  public async initializeJobQueue(): Promise<void>
+  {
+    // set all jobs that are currently running to ABORTED
+    await this._resetAllRunningJobs();
+
+    setTimeout(this._jobLoop.bind(this), INTERVAL - new Date().getTime() % INTERVAL);
   }
 
   public async pause(id: number): Promise<JobConfig[]>
@@ -223,7 +244,7 @@ export class JobQueue
       try
       {
         this.runningJobs.get(id).pause();
-        await this._setJobStatus(id, true, 'PAUSED');
+        await this.setJobStatus(id, true, 'PAUSED');
         return resolve(await this.get(id) as JobConfig[]);
       }
       catch (e)
@@ -244,7 +265,7 @@ export class JobQueue
         return reject(new Error('Job not found.'));
       }
 
-      await this._setJobStatus(id, true, 'RUNNING');
+      await this.setJobStatus(id, true, 'RUNNING');
       getJobs[0] = await this._setRunNow(getJobs[0]);
       return resolve(await App.DB.upsert(this.jobTable, getJobs[0]) as JobConfig[]);
     });
@@ -265,7 +286,7 @@ export class JobQueue
       {
         return reject(new Error('Too many jobs set to run now currently in the queue.'));
       }
-      const status: boolean = await this._setJobStatus(getJobs[0].id, true, 'RUNNING');
+      const status: boolean = await this.setJobStatus(getJobs[0].id, true, 'RUNNING');
       if (!status)
       {
         winston.warn('Job running status was not toggled.');
@@ -304,11 +325,44 @@ export class JobQueue
       const jobResult: TaskOutputConfig = await this.runningRunNowJobs.get(getJobs[0].id).run() as TaskOutputConfig;
       const jobsFromId: JobConfig[] = await this.get(getJobs[0].id);
       const jobStatus: string = jobResult.status === true ? 'SUCCESS' : 'FAILURE';
-      await this._setJobStatus(jobsFromId[0].id, false, jobStatus);
+      await this.setJobStatus(jobsFromId[0].id, false, jobStatus);
       this.runningRunNowJobs.delete(getJobs[0].id);
-      // TODO: log job result
 
-      return resolve(jobResult['options']['outputStream'] as stream.Readable);
+      // log job result
+      const jobLogConfig: JobLogConfig[] = await App.JobL.create(getJobs[0].id, jobResult['options']['logStream']);
+      await this._setJobLogId(getJobs[0].id, jobLogConfig[0].id);
+      return resolve(jobResult.options.outputStream as stream.Readable);
+    });
+  }
+
+  // Status codes: PENDING SUCCESS FAILURE PAUSED CANCELED RUNNING ABORTED (PAUSED/RUNNING when midway was restarted)
+  public async setJobStatus(id: number, running: boolean, status: string): Promise<boolean>
+  {
+    return new Promise<boolean>(async (resolve, reject) =>
+    {
+      const jobs: JobConfig[] = await this._select([], { id }) as JobConfig[];
+      if (jobs.length === 0)
+      {
+        return resolve(false);
+      }
+      if (jobs[0].running === running)
+      {
+        return resolve(false);
+      }
+
+      if (jobs[0].running === false && running === true) // set start time
+      {
+        jobs[0].startTime = new Date();
+      }
+      if (jobs[0].running === true && running === false) // set end time
+      {
+        jobs[0].endTime = new Date();
+      }
+
+      jobs[0].running = running;
+      jobs[0].status = status;
+      const doNothing: JobConfig[] = await App.DB.upsert(this.jobTable, jobs[0]) as JobConfig[];
+      resolve(true);
     });
   }
 
@@ -321,7 +375,7 @@ export class JobQueue
         if (this.runningJobs.has(id))
         {
           // set job status back to RUNNING
-          await this._setJobStatus(id, true, 'RUNNING');
+          await this.setJobStatus(id, true, 'RUNNING');
 
           // resolve updated job so that we can continue execution without blocking on the response
           resolve(await this.get(id) as JobConfig[]);
@@ -330,9 +384,12 @@ export class JobQueue
           const jobResult: TaskOutputConfig = await this.runningJobs.get(id).run() as TaskOutputConfig;
           const jobsFromId: JobConfig[] = await this.get(id);
           const jobStatus: string = jobResult.status === true ? 'SUCCESS' : 'FAILURE';
-          await this._setJobStatus(id, false, jobStatus);
+          await this.setJobStatus(id, false, jobStatus);
           await App.SKDR.setRunning(jobsFromId[0].scheduleId, false);
           this.runningJobs.delete(id);
+
+          const jobLogConfig: JobLogConfig[] = await App.JobL.create(id, jobResult['options']['logStream']);
+          await this._setJobLogId(id, jobLogConfig[0].id);
         }
       }
       catch (e)
@@ -341,15 +398,6 @@ export class JobQueue
       }
       return reject(new Error('Job not found.'));
     });
-
-  }
-
-  public async initializeJobQueue(): Promise<void>
-  {
-    // set all jobs that are currently running to ABORTED
-    await this._resetAllRunningJobs();
-
-    setTimeout(this._jobLoop.bind(this), INTERVAL - new Date().getTime() % INTERVAL);
   }
 
   private async _checkJobTable(): Promise<void>
@@ -365,11 +413,11 @@ export class JobQueue
       }
       const query = new Tasty.Query(this.jobTable).filter(this.jobTable['status'].equals('PENDING'))
         .filter(this.jobTable['priority'].greaterThan(-1))
-        .filter(this.jobTable['running'].equals('false')).sort(this.jobTable['priority'], 'asc')
+        .filter(this.jobTable['running'].equals(false)).sort(this.jobTable['priority'], 'asc')
         .sort(this.jobTable['runNowPriority'], 'desc').sort(this.jobTable['createdAt'], 'asc').take(newJobSlots);
       const queryStr: string = App.DB.getDB().generateString(query);
       const rawResults = await App.DB.getDB().execute([queryStr]);
-      const jobs: JobConfig[] = rawResults.map((result: object) => new JobConfig(result as JobConfig));
+      const jobs: JobConfig[] = rawResults.map((result) => new JobConfig(result as JobConfig));
 
       let i = 0;
       while (i < newJobSlots)
@@ -402,7 +450,7 @@ export class JobQueue
           // update the table to running = true
 
           this.runningJobs.set(nextJob.id, newJob);
-          const status: boolean = await this._setJobStatus(nextJob.id, true, 'RUNNING');
+          const status: boolean = await this.setJobStatus(nextJob.id, true, 'RUNNING');
           if (!status)
           {
             winston.warn('Job running status was not toggled.');
@@ -418,11 +466,13 @@ export class JobQueue
         const jobResult: TaskOutputConfig = await this.runningJobs.get(jobId).run() as TaskOutputConfig;
         const jobsFromId: JobConfig[] = await this.get(jobId);
         const jobStatus: string = jobResult.status === true ? 'SUCCESS' : 'FAILURE';
-        await this._setJobStatus(jobsFromId[0].id, false, jobStatus);
+        await this.setJobStatus(jobsFromId[0].id, false, jobStatus);
         await App.SKDR.setRunning(jobsFromId[0].scheduleId, false);
         this.runningJobs.delete(jobId);
-        // TODO: log job result
         winston.info('Job result: ' + JSON.stringify(jobResult, null, 2));
+
+        const jobLogConfig: JobLogConfig[] = await App.JobL.create(jobId, jobResult['options']['logStream']);
+        await this._setJobLogId(jobId, jobLogConfig[0].id);
       });
 
     });
@@ -476,8 +526,7 @@ export class JobQueue
     });
   }
 
-  // Status codes: SUCCESS FAILURE PAUSED CANCELED RUNNING ABORTED (PAUSED/RUNNING when midway was restarted)
-  private async _setJobStatus(id: number, running: boolean, status: string): Promise<boolean>
+  private async _setJobLogId(id: number, jobLogId: number): Promise<boolean>
   {
     return new Promise<boolean>(async (resolve, reject) =>
     {
@@ -486,12 +535,8 @@ export class JobQueue
       {
         return resolve(false);
       }
-      if (jobs[0].running === running)
-      {
-        return resolve(false);
-      }
-      jobs[0].running = running;
-      jobs[0].status = status;
+
+      jobs[0].logId = jobLogId;
       const doNothing: JobConfig[] = await App.DB.upsert(this.jobTable, jobs[0]) as JobConfig[];
       resolve(true);
     });
@@ -504,12 +549,12 @@ export class JobQueue
     {
       let maxRunNowPriority: number = 1;
       const query = new Tasty.Query(this.jobTable).filter(this.jobTable['status'].equals('PENDING'))
-        .filter(this.jobTable['running'].equals('false')).filter(this.jobTable['priority'].equals(0))
+        .filter(this.jobTable['running'].equals(false)).filter(this.jobTable['priority'].equals(0))
         .sort(this.jobTable['runNowPriority'], 'desc').take(1);
       const queryStr: string = App.DB.getDB().generateString(query);
       const rawResults = await App.DB.getDB().execute([queryStr]);
 
-      const jobs: JobConfig[] = rawResults.map((result: object) => new JobConfig(result));
+      const jobs: JobConfig[] = rawResults.map((result: object) => new JobConfig(result as JobConfig));
       if (jobs.length !== 0)
       {
         maxRunNowPriority = jobs[0].runNowPriority;
