@@ -51,6 +51,7 @@ const { List, Map } = Immutable;
 
 import { FileConfig, SinkConfig, SourceConfig } from 'shared/etl/immutable/EndpointRecords';
 import { ETLTemplate, SinksMap, SourcesMap } from 'shared/etl/immutable/TemplateRecords';
+import LanguageController from 'shared/etl/languages/LanguageControllers';
 import { Sinks, Sources } from 'shared/etl/types/EndpointTypes';
 import { FieldTypes, Languages } from 'shared/etl/types/ETLTypes';
 import { TransformationEngine } from 'shared/transformations/TransformationEngine';
@@ -247,13 +248,9 @@ export class TemplateProxy
       EngineUtil.addInitialTypeCasts(engine);
 
       const language = this.template.getEdgeLanguage(edgeId);
-      switch (language)
+      if (language !== Languages.JavaScript)
       {
-        case 'elastic':
-          EngineUtil.autodetectElasticTypes(engine, documentConfig.documents);
-          break;
-        default:
-          break;
+        LanguageController.get(language).autodetectFieldTypes(engine, documentConfig.documents);
       }
     }
     else
