@@ -79,7 +79,7 @@ export const SchedulableSinks: Sinks[] =
   [Sinks.Database, Sinks.Sftp, Sinks.Http, Sinks.Fs];
 
 export const SchedulableSources: Sources[] =
-  [Sources.Algorithm, Sources.Sftp, Sources.Http, Sources.Fs];
+  [Sources.Algorithm, Sources.Sftp, Sources.Http, Sources.Fs, Sources.Mysql, Sources.Postgresql];
 
 export interface SourceConfig
 {
@@ -144,9 +144,9 @@ export const SourceOptionsDefaults: SourceOptionsTypes =
     Http: {
       method: 'GET',
       headers: {
-        accept: '',
         contentType: 'application/json',
       },
+      params: {},
     },
     Fs: {
       path: '',
@@ -193,9 +193,9 @@ export const SinkOptionsDefaults: SinkOptionsTypes =
     Http: {
       method: 'POST',
       headers: {
-        accept: '',
         contentType: 'application/json',
       },
+      params: {},
     },
     Fs: {
       path: '',
@@ -213,8 +213,12 @@ export interface HttpOptions
 {
   method: 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
   headers: {
-    accept: string;
-    contentType: string;
+    contentType?: string;
+    accept?: string;
+    [k: string]: any;
+  };
+  params: {
+    [k: string]: any;
   };
 }
 
