@@ -44,21 +44,28 @@ THE SOFTWARE.
 
 // Copyright 2018 Terrain Data, Inc.
 
-import AddTransformationNode from 'shared/transformations/nodes/AddTransformationNode';
-import ArraySumTransformationNode from 'shared/transformations/nodes/ArraySumTransformationNode';
-import DivideTransformationNode from 'shared/transformations/nodes/DivideTransformationNode';
-import FindReplaceTransformationNode from 'shared/transformations/nodes/FindReplaceTransformationNode';
-import HashTransformationNode from 'shared/transformations/nodes/HashTransformationNode';
-import MultiplyTransformationNode from 'shared/transformations/nodes/MultiplyTransformationNode';
-import SetIfTransformationNode from 'shared/transformations/nodes/SetIfTransformationNode';
-import SubtractTransformationNode from 'shared/transformations/nodes/SubtractTransformationNode';
+import AddTransformationNode from './nodes/AddTransformationNode';
+import ArrayCountTransformationNode from './nodes/ArrayCountTransformationNode';
+import ArraySumTransformationNode from './nodes/ArraySumTransformationNode';
 import CastTransformationNode from './nodes/CastTransformationNode';
+import DecryptTransformationNode from './nodes/DecryptTransformationNode';
+import DifferenceTransformationNode from './nodes/DifferenceTransformationNode';
+import DivideTransformationNode from './nodes/DivideTransformationNode';
 import DuplicateTransformationNode from './nodes/DuplicateTransformationNode';
+import EncryptTransformationNode from './nodes/EncryptTransformationNode';
 import FilterTransformationNode from './nodes/FilterTransformationNode';
+import FindReplaceTransformationNode from './nodes/FindReplaceTransformationNode';
+import HashTransformationNode from './nodes/HashTransformationNode';
 import InsertTransformationNode from './nodes/InsertTransformationNode';
 import JoinTransformationNode from './nodes/JoinTransformationNode';
+import MultiplyTransformationNode from './nodes/MultiplyTransformationNode';
+import ProductTransformationNode from './nodes/ProductTransformationNode';
+import QuotientTransformationNode from './nodes/QuotientTransformationNode';
+import SetIfTransformationNode from './nodes/SetIfTransformationNode';
 import SplitTransformationNode from './nodes/SplitTransformationNode';
 import SubstringTransformationNode from './nodes/SubstringTransformationNode';
+import SubtractTransformationNode from './nodes/SubtractTransformationNode';
+import SumTransformationNode from './nodes/SumTransformationNode';
 import TransformationNode from './nodes/TransformationNode';
 import UppercaseTransformationNode from './nodes/UppercaseTransformationNode';
 import { TransformationEngine } from './TransformationEngine';
@@ -177,11 +184,30 @@ const TransformationNodeInfo: AllNodeInfoType =
       },
     [TransformationNodeType.InsertNode]:
       {
-        humanName: 'Insert String',
+        humanName: 'Append / Prepend',
         editable: true,
         creatable: true,
-        description: `Insert a string at a given position in this field's value`,
+        description: `Append, Prepend, or Insert Text`,
         type: InsertTransformationNode,
+        isAvailable: (engine, fieldId) =>
+        {
+          return EngineUtil.getRepresentedType(fieldId, engine) === 'string';
+        },
+        shortSummary: (meta) =>
+        {
+          if (meta.at === -1)
+          {
+            return 'Append Text';
+          }
+          else if (meta.at === 0)
+          {
+            return 'Prepend Text';
+          }
+          else
+          {
+            return `Insert Text at Position ${meta.at}`;
+          }
+        },
         targetedVisitor: (visitor: TransformationNodeVisitor,
           transformationNode: TransformationNode,
           docCopy: object,
@@ -214,6 +240,10 @@ const TransformationNodeInfo: AllNodeInfoType =
         isAvailable: (engine, fieldId) =>
         {
           return EngineUtil.getRepresentedType(fieldId, engine) === 'string';
+        },
+        shortSummary: (meta) =>
+        {
+          return `Substring from ${meta.from} to ${meta.from + meta.length}`;
         },
         type: SubstringTransformationNode,
         targetedVisitor: (visitor: TransformationNodeVisitor,
@@ -260,28 +290,6 @@ const TransformationNodeInfo: AllNodeInfoType =
           options: object) =>
           visitor.visitHashNode(transformationNode, docCopy, options),
       },
-    [TransformationNodeType.ArraySumNode]:
-      {
-        humanName: 'Array Sum',
-        editable: false,
-        creatable: true,
-        description: `Sums the values of this array`,
-        type: ArraySumTransformationNode,
-        isAvailable: (engine, fieldId) =>
-        {
-          return (
-            EngineUtil.getRepresentedType(fieldId, engine) === 'array' &&
-            EngineUtil.getValueType(fieldId, engine) === 'number' &&
-            EngineUtil.isNamedField(engine.getOutputKeyPath(fieldId))
-          );
-        },
-        targetedVisitor: (visitor: TransformationNodeVisitor,
-          transformationNode: TransformationNode,
-          docCopy: object,
-          options: object) =>
-          visitor.visitArraySumNode(transformationNode, docCopy, options),
-        newFieldType: 'number',
-      },
     [TransformationNodeType.AddNode]:
       {
         humanName: 'Add',
@@ -291,6 +299,10 @@ const TransformationNodeInfo: AllNodeInfoType =
         isAvailable: (engine, fieldId) =>
         {
           return EngineUtil.getRepresentedType(fieldId, engine) === 'number';
+        },
+        shortSummary: (meta) =>
+        {
+          return `Add ${meta.shift}`;
         },
         type: AddTransformationNode,
         targetedVisitor: (visitor: TransformationNodeVisitor,
@@ -309,6 +321,10 @@ const TransformationNodeInfo: AllNodeInfoType =
         {
           return EngineUtil.getRepresentedType(fieldId, engine) === 'number';
         },
+        shortSummary: (meta) =>
+        {
+          return `Subtract ${meta.shift}`;
+        },
         type: SubtractTransformationNode,
         targetedVisitor: (visitor: TransformationNodeVisitor,
           transformationNode: TransformationNode,
@@ -326,6 +342,10 @@ const TransformationNodeInfo: AllNodeInfoType =
         {
           return EngineUtil.getRepresentedType(fieldId, engine) === 'number';
         },
+        shortSummary: (meta) =>
+        {
+          return `Multiply by ${meta.factor}`;
+        },
         type: MultiplyTransformationNode,
         targetedVisitor: (visitor: TransformationNodeVisitor,
           transformationNode: TransformationNode,
@@ -342,6 +362,10 @@ const TransformationNodeInfo: AllNodeInfoType =
         isAvailable: (engine, fieldId) =>
         {
           return EngineUtil.getRepresentedType(fieldId, engine) === 'number';
+        },
+        shortSummary: (meta) =>
+        {
+          return `Divide by ${meta.factor}`;
         },
         type: DivideTransformationNode,
         targetedVisitor: (visitor: TransformationNodeVisitor,
@@ -385,6 +409,168 @@ const TransformationNodeInfo: AllNodeInfoType =
           docCopy: object,
           options: object) =>
           visitor.visitFindReplaceNode(transformationNode, docCopy, options),
+      },
+    [TransformationNodeType.ArraySumNode]:
+      {
+        humanName: 'Array Sum',
+        editable: false,
+        creatable: true,
+        description: `Sum the entries of an array`,
+        type: ArraySumTransformationNode,
+        isAvailable: (engine, fieldId) =>
+        {
+          return (
+            EngineUtil.getRepresentedType(fieldId, engine) === 'array' &&
+            EngineUtil.getValueType(fieldId, engine) === 'number' &&
+            EngineUtil.isNamedField(engine.getOutputKeyPath(fieldId))
+          );
+        },
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitArraySumNode(transformationNode, docCopy, options),
+        newFieldType: 'number',
+      },
+    [TransformationNodeType.ArrayCountNode]:
+      {
+        humanName: 'Array Count',
+        editable: false,
+        creatable: true,
+        description: `Counts how many elements are in an array`,
+        type: ArrayCountTransformationNode,
+        isAvailable: (engine, fieldId) =>
+        {
+          return (
+            EngineUtil.getRepresentedType(fieldId, engine) === 'array' &&
+            EngineUtil.getValueType(fieldId, engine) === 'number' &&
+            EngineUtil.isNamedField(engine.getOutputKeyPath(fieldId))
+          );
+        },
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitArrayCountNode(transformationNode, docCopy, options),
+        newFieldType: 'number',
+      },
+    [TransformationNodeType.ProductNode]:
+      {
+        humanName: 'Product of Fields',
+        editable: false,
+        creatable: true,
+        description: `Multiplies two or more fields together and puts the result in a new field`,
+        type: ProductTransformationNode,
+        isAvailable: (engine, fieldId) =>
+        {
+          return (
+            EngineUtil.getRepresentedType(fieldId, engine) === 'number' &&
+            EngineUtil.isNamedField(engine.getOutputKeyPath(fieldId))
+          );
+        },
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitProductNode(transformationNode, docCopy, options),
+        newFieldType: 'number',
+      },
+    [TransformationNodeType.QuotientNode]:
+      {
+        humanName: 'Quotient of Fields',
+        editable: false,
+        creatable: true,
+        description: `Divides two fields and puts the result in a new field`,
+        type: QuotientTransformationNode,
+        isAvailable: (engine, fieldId) =>
+        {
+          return (
+            EngineUtil.getRepresentedType(fieldId, engine) === 'number' &&
+            EngineUtil.isNamedField(engine.getOutputKeyPath(fieldId))
+          );
+        },
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitQuotientNode(transformationNode, docCopy, options),
+        newFieldType: 'number',
+      },
+    [TransformationNodeType.SumNode]:
+      {
+        humanName: 'Sum of Fields',
+        editable: false,
+        creatable: true,
+        description: `Sums two or more fields and puts the result in a new field`,
+        type: SumTransformationNode,
+        isAvailable: (engine, fieldId) =>
+        {
+          return (
+            EngineUtil.getRepresentedType(fieldId, engine) === 'number' &&
+            EngineUtil.isNamedField(engine.getOutputKeyPath(fieldId))
+          );
+        },
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitSumNode(transformationNode, docCopy, options),
+        newFieldType: 'number',
+      },
+    [TransformationNodeType.DifferenceNode]:
+      {
+        humanName: 'Difference of Fields',
+        editable: false,
+        creatable: true,
+        description: `Subtracts one field from another and puts the result in a new field`,
+        type: DifferenceTransformationNode,
+        isAvailable: (engine, fieldId) =>
+        {
+          return (
+            EngineUtil.getRepresentedType(fieldId, engine) === 'number' &&
+            EngineUtil.isNamedField(engine.getOutputKeyPath(fieldId))
+          );
+        },
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitDifferenceNode(transformationNode, docCopy, options),
+        newFieldType: 'number',
+      },
+    [TransformationNodeType.EncryptNode]:
+      {
+        humanName: 'Encrypt',
+        editable: true,
+        creatable: true,
+        description: `Encrypt a field using the secure AES algorithm`,
+        isAvailable: (engine, fieldId) =>
+        {
+          return EngineUtil.getRepresentedType(fieldId, engine) === 'string';
+        },
+        type: EncryptTransformationNode,
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitEncryptNode((transformationNode as EncryptTransformationNode), docCopy, options),
+      },
+    [TransformationNodeType.DecryptNode]:
+      {
+        humanName: 'Decrypt',
+        editable: true,
+        creatable: true,
+        description: `Decrypt a field that was previously encrypted with an Encrypt transformation`,
+        isAvailable: (engine, fieldId) =>
+        {
+          return EngineUtil.getRepresentedType(fieldId, engine) === 'string';
+        },
+        type: DecryptTransformationNode,
+        targetedVisitor: (visitor: TransformationNodeVisitor,
+          transformationNode: TransformationNode,
+          docCopy: object,
+          options: object) =>
+          visitor.visitDecryptNode((transformationNode as DecryptTransformationNode), docCopy, options),
       },
   };
 
