@@ -43,14 +43,17 @@ THE SOFTWARE.
 */
 
 // Copyright 2018 Terrain Data, Inc.
+import ETLRouteUtil from 'app/etl/ETLRouteUtil';
 import { ModalProps, MultiModal } from 'common/components/overlay/MultiModal';
 import TerrainComponent from 'common/components/TerrainComponent';
 import TerrainTabs from 'common/components/TerrainTabs';
 import { ETLActions } from 'etl/ETLRedux';
 import IntegrationList from 'etl/integrations/components/IntegrationList';
 import TemplateList from 'etl/templates/components/TemplateList';
+import Jobs from 'jobs/components/Jobs';
 import * as React from 'react';
 import { Tab, TabList, TabPanel, Tabs } from 'react-tabs';
+import ScheduleList from 'scheduler/components/ScheduleList';
 import { ETLTemplate } from 'shared/etl/immutable/TemplateRecords';
 import Util from 'util/Util';
 import './DataTabs.less';
@@ -70,17 +73,19 @@ class DataTabs extends TerrainComponent<DataTabsProps>
   public tabs = [
     { key: 'templates', label: 'Templates' },
     { key: 'integrations', label: 'Integrations' },
-    { key: 'runnow', label: 'Import / Export Run Now' },
+    { key: 'schedules', label: 'Schedules' },
+    { key: 'jobs', label: 'Jobs' },
+
+    { key: 'newtemplate', label: 'New Import or Export' },
   ];
 
   public tabToRouteMap = {
     templates: '/data/templates',
     integrations: '/data/integrations',
-    runnow: '/data/runnow',
-  };
+    schedules: '/data/schedules',
+    jobs: '/data/jobs',
 
-  public state = {
-    tabIndex: this.tabs.indexOf(this.props.params.tab),
+    newtemplate: '/data/newtemplate',
   };
 
   public componentDidMount()
@@ -115,7 +120,9 @@ class DataTabs extends TerrainComponent<DataTabsProps>
             tabToRouteMap={this.tabToRouteMap}
             router={router}
           >
-            {this.props.children}
+            {
+              this.props.children
+            }
           </TerrainTabs>
           <MultiModal
             requests={modalRequests}
