@@ -60,6 +60,8 @@ const { List, Map } = Immutable;
 import Menu from 'common/components/Menu';
 import { tooltip } from 'common/components/tooltip/Tooltips';
 import { TemplateField } from 'etl/templates/FieldTypes';
+import LanguageUI from 'etl/templates/languages/LanguageUI';
+import EngineUtil from 'shared/transformations/util/EngineUtil';
 import { instanceFnDecorator } from 'shared/util/Classes';
 import { mapDispatchKeys, mapStateKeys, TemplateEditorField, TemplateEditorFieldProps } from './TemplateEditorField';
 
@@ -136,6 +138,12 @@ class EditorFieldPreview extends TemplateEditorField<Props>
     );
   }
 
+  public isPrimaryKey()
+  {
+    const language = this._getCurrentLanguage();
+    return LanguageUI.get(language).isFieldPrimaryKey(this._field().fieldProps);
+  }
+
   public render()
   {
     const { canEdit, preview, labelOverride, labelOnly } = this.props;
@@ -173,18 +181,23 @@ class EditorFieldPreview extends TemplateEditorField<Props>
             onMouseLeave={this.handleMouseLeave}
             style={labelStyle}
           >
+            {
+              this.renderTypeIcon()
+            }
             <div
               className={classNames({
                 'field-preview-label': true,
                 'field-preview-array-label': field.isWildcardField(),
                 'field-preview-can-toggle': this.props.toggleOpen !== undefined,
               })}
-              onClick={this.props.toggleOpen}
+              onClick={this.handleLabelClicked}
+              style={fontColor(Colors().text1, Colors().active)}
+              key='label'
             >
               {labelOverride != null ? labelOverride : field.name}
             </div>
             {
-              this._isPrimaryKey() ?
+              this.isPrimaryKey() ?
                 <div
                   className='primary-key-icon'
                   style={fontColor(Colors().active)}
@@ -333,7 +346,39 @@ class EditorFieldPreview extends TemplateEditorField<Props>
       this.openSettings();
     }
   }
+
+  private renderTypeIcon()
+  {
+    const type = EngineUtil.getRepresentedType(this.props.fieldId, this._currentEngine());
+    const Icon = typeToIcon[type];
+    return (
+      <div
+        className='field-type-icon'
+        style={fontColor(Colors().text2)}
+      >
+        <Icon />
+      </div>
+    );
+  }
 }
+
+const TextTypeIcon = require('./../../../../../images/icon_type_text.svg?name=TextTypeIcon');
+const DateTypeIcon = require('./../../../../../images/icon_type_date.svg?name=DateTypeIcon');
+const NumberTypeIcon = require('./../../../../../images/icon_type_number.svg?name=NumberTypeIcon');
+const ArrayTypeIcon = require('./../../../../../images/icon_type_array.svg?name=ArrayTypeIcon');
+const ObjectTypeIcon = require('./../../../../../images/icon_type_object.svg?name=ObjectTypeIcon');
+const GeoTypeIcon = require('./../../../../../images/icon_type_geo.svg?name=GeoTypeIcon');
+const BooleanTypeIcon = require('./../../../../../images/icon_type_boolean.svg?name=BooleanTypeIcon');
+
+const typeToIcon = {
+  string: TextTypeIcon,
+  object: ObjectTypeIcon,
+  number: NumberTypeIcon,
+  boolean: BooleanTypeIcon,
+  array: ArrayTypeIcon,
+  date: DateTypeIcon,
+  geo: GeoTypeIcon,
+};
 
 const emptyOptions = List([]);
 
