@@ -46,7 +46,6 @@ THE SOFTWARE.
 // tslint:disable:no-console strict-boolean-expressions no-var-requires
 import PathfinderCreateLine from 'app/builder/components/pathfinder/PathfinderCreateLine';
 import Colors, { backgroundColor, borderColor, fontColor, getStyle } from 'app/colors/Colors';
-import Modal from 'app/common/components/Modal';
 import { ETLActions } from 'app/etl/ETLRedux';
 import { ETLState } from 'app/etl/ETLTypes';
 import { SchedulerActions } from 'app/scheduler/data/SchedulerRedux';
@@ -82,10 +81,8 @@ class ScheduleList extends TerrainComponent<Props>
 
   public state: {
     confirmModalOpen: boolean,
-    deleteScheduleId: ID,
   } = {
       confirmModalOpen: false,
-      deleteScheduleId: -1,
     };
 
   public componentWillMount()
@@ -138,20 +135,11 @@ class ScheduleList extends TerrainComponent<Props>
     });
   }
 
-  public performAction(action, scheduleId?: ID)
+  public performAction(action, scheduleId: ID)
   {
-    scheduleId = scheduleId !== undefined ? scheduleId : this.state.deleteScheduleId;
     this.props.schedulerActions({
       actionType: action,
       scheduleId,
-    });
-  }
-
-  public confirmDeleteSchedule(scheduleId: ID)
-  {
-    this.setState({
-      confirmModalOpen: true,
-      deleteScheduleId: scheduleId,
     });
   }
 
@@ -160,20 +148,15 @@ class ScheduleList extends TerrainComponent<Props>
     const { schedules } = this.props;
     const keys = schedules.keySeq().toList().sort();
     const scheduleList = keys.map((id) => schedules.get(id));
-    const toDelete = schedules.get(this.state.deleteScheduleId) ? schedules.get(this.state.deleteScheduleId).name : '';
+
     return (
       <div
         className='schedule-list-wrapper'
-        style={_.extend(backgroundColor(Colors().blockBg), borderColor(Colors().blockOutline))}
+        style={backgroundColor(Colors().blockBg)}
       >
         <div
           className='schedule-list-header'
         >
-          <div
-            className='schedule-list-title'
-          >
-            Schedules
-          </div>
           <div
             className='schedule-list-refresh-wrapper'
           >
@@ -193,9 +176,9 @@ class ScheduleList extends TerrainComponent<Props>
           {
             scheduleList.map((schedule, i) =>
               <Schedule
-                key={i}
+                key={schedule.id}
                 schedule={schedule}
-                onDelete={this.confirmDeleteSchedule}
+                onDelete={this._fn(this.performAction, 'deleteSchedule')}
                 onRun={this._fn(this.performAction, 'runSchedule')}
                 onPause={this._fn(this.performAction, 'pauseSchedule')}
                 onUnpause={this._fn(this.performAction, 'unpauseSchedule')}
@@ -213,15 +196,6 @@ class ScheduleList extends TerrainComponent<Props>
           canEdit={TerrainTools.isAdmin()}
           onCreate={this.createSchedule}
           showText={true}
-        />
-        <Modal
-          open={this.state.confirmModalOpen}
-          confirm={true}
-          title={'Confirm Action'}
-          message={`Are you sure you want to delete ${toDelete}?`}
-          confirmButtonText={'Yes'}
-          onConfirm={this._fn(this.performAction, 'deleteSchedule')}
-          onClose={this._toggle('confirmModalOpen')}
         />
       </div>
     );
